moves job routes to its own router
